Fixed error message when accessing with /bluedragon/administrator/
Accessing the Administration with /bluedragon/administrator/ only threw an error of:
```
Method getRootURL could not be found. Check you have correct method name, the method name casing matches that of the Java class and you've provided the correct number of arguments
```

The method obviously does not exists. I've fixed the code to work with the getContextPath(). Tested on MacOS X and Linux servers.